pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "sapho"
version = "0.1.0"
description = "Deep Research Agent for Crypto"
readme = "README.md"
requires-python = ">=3.9"
license = "MIT"
authors = [
    { name = "Operator Labs", email = "support@operator.io" }
]
dependencies = [
    "fastapi>=0.104.1",
    "uvicorn>=0.24.0",
    "python-dotenv>=1.0.0",
    "openai>=1.3.7",
    "aiohttp>=3.9.1",
    "pydantic>=2.5.2",
    "python-multipart>=0.0.6"
]
